(Wish You Were Here) Minor changes to work with new set_cost hook
Also made it so that its initial sell value is calculated when the joker
is added to the deck, since the original set_cost hook in charge of this
functionality was massively changed to account for Hole in One's
implementation (Which also benefitted the stability of Wish You Were
Here). Hopefully there are no major issues with mod compatibility with
this update.