(Basic Joker Energy) Check if there's room before copying + small tweaks
In the current version of steamodded the message will show on blueprint twice, but it'll be fixed in a later version, so this is okay

SMODS.eval_this is deprecated in the latest Steamodded version, and the
closest alternative to this function is the new SMODS.calculate_effect
function. This does mean that the minimum steamodded version to run the
mod will need to be updated to reflect this change.